+++ src/main/java/com/saintdan/util/rocketmq/param/ContactsParam.java
@@ -5,6 +5,8 @@ import java.io.Serializable;
 /**
  * 联系人参数
  *
+ * TODO ONS管理平台中,可创建多个联系人,现在拼接方法不明.可能会有变化.
+ *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 1/20/16
  * @since JDK1.8
@@ -16,6 +18,15 @@ public class ContactsParam implements Serializable {
     private String name;
 
     private String phone;
+
+    public ContactsParam() {
+
+    }
+
+    public ContactsParam(String name, String phone) {
+        this.name = name;
+        this.phone = phone;
+    }
 
     public String getName() {
         return name;